break out zenerdiode
